refactor: add monaco editor default options and setGlobalEditorTheme api
