+++ packages/plugins/schema/src/Main.vue
@@ -43,7 +43,6 @@ import { nextTick, reactive, getCurrentInstance, onActivated, ref } from 'vue'
 import { Popover } from '@opentiny/vue'
 import { VueMonaco, CloseIcon } from '@opentiny/tiny-engine-common'
 import { useCanvas, useModal, useHistory, useNotify } from '@opentiny/tiny-engine-controller'
-import { theme } from '@opentiny/tiny-engine-controller/js/monaco'
 import { utils } from '@opentiny/tiny-engine-utils'
 import { iconSave, iconDownloadLink } from '@opentiny/vue-icon'
 
@@ -132,15 +131,9 @@ export default {
       close,
       showRed,
       options: {
-        roundedSelection: true,
-        automaticLayout: true,
-        autoIndent: true,
         language: 'json',
         // readOnly: !pageState.isLock,  暂时放开schema录入功能，等画布功能完善后，再打开此注释
         readOnly: false,
-        formatOnPaste: true,
-        tabSize: 2,
-        theme: theme(),
         minimap: {
           enabled: false
         }
