+++ packages/settings/events/src/components/BindEventsDialog.vue
@@ -89,7 +89,6 @@ import { reactive, ref, watchEffect, nextTick } from 'vue'
 import { VueMonaco } from '@opentiny/tiny-engine-common'
 import { Button, DialogBox, Input, Search, Popover, Switch } from '@opentiny/vue'
 import { useCanvas, useHistory, useLayout } from '@opentiny/tiny-engine-controller'
-import { theme } from '@opentiny/tiny-engine-controller/js/monaco'
 import { string2Ast, ast2String } from '@opentiny/tiny-engine-controller/js/ast'
 import { iconYes, iconHelpQuery } from '@opentiny/vue-icon'
 
@@ -152,13 +151,7 @@ export default {
     })
 
     const editorOptions = {
-      roundedSelection: true,
-      automaticLayout: true,
-      autoIndent: true,
       language: 'json',
-      formatOnPaste: true,
-      tabSize: 2,
-      theme: theme(),
       lineNumbers: false,
       minimap: {
         enabled: false
